+++ isoft-fm-foundation.php
@@ -164,6 +164,7 @@ add_action(
 			( new ISOFT_FMF_License_Manager() )->register_hooks();
 			( new ISOFT_FMF_Licenses_Page() )->register_hooks();
 			( new ISOFT_FMF_Stats_Page() )->register_hooks();
+			( new ISOFT_FMF_Log_Page() )->register_hooks();
 			( new ISOFT_FMF_Pdf_Thumbnail() )->register_hooks();
 			( new ISOFT_FMF_Tinymce() )->register_hooks();
 			( new ISOFT_FMF_Demo_Content() )->register_hooks();